+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Localize wikidata interests.
 - Add popover to interests.
 - Add lint-staged with husky pre-commit hook.
+- Add i18n to profile description
 
 ### Changed